requirements.txt
pymongo
flask
flask-cors
pandas==2.3.3
scikit-learn==1.7.2
xgboost==3.1.1
requests
python-dotenv